download kami from releases

+++ src/main/java/me/zeroeightsix/installer/Installer.java
@@ -1,17 +1,14 @@
 package me.zeroeightsix.installer;
 
 
-import java.io.File;
+import java.io.*;
 import javax.imageio.ImageIO;
 import javax.swing.*;
-import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 import java.awt.*;
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.MalformedURLException;
 import java.net.URL;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -19,10 +16,6 @@ import java.nio.file.StandardCopyOption;
 import java.util.Locale;
 import java.util.Random;
 
-import net.minecraft.util.Util.*;
-import org.w3c.dom.Document;
-import org.w3c.dom.Node;
-import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
 
 public class Installer extends JPanel {
@@ -79,16 +72,14 @@ public class Installer extends JPanel {
         kamiIcon.setBounds(236, 70, 128, 128);
         backgroundPane.setBounds(0, 0, 600, 355);
 
-        installButton.addActionListener(e -> {
+        installButton.addActionListener(event -> {
             installButton.setEnabled(false);
             installIcon.setOpaque(false);
-
             try {
-                InstallKami(getModsFolder());
-            } catch (IOException | SAXException | ParserConfigurationException Exception) {
-                Exception.printStackTrace();
+                installKami(getModsFolder());
+            } catch (IOException | SAXException | ParserConfigurationException e) {
+                e.printStackTrace();
             }
-
             System.exit(0);
         });
     }
@@ -114,6 +105,7 @@ public class Installer extends JPanel {
         if (operatingSystemName.contains("nux")) {
             return System.getProperty("user.home") + "/.local/share/multimc/instances";
         } else if (operatingSystemName.contains("mac") || operatingSystemName.contains("darwin")) {
+            //TODO: get instances folder for osx
             return "TODO Mac multimc instances folder";
         } else if (operatingSystemName.contains("win")) {
             return System.getenv("PROGRAMFILES") + File.separator + "MultiMC" + File.separator + "instances" + File.separator;
@@ -122,32 +114,35 @@ public class Installer extends JPanel {
         throw new RuntimeException("Cannot find instaces folder.");
     }
 
-    private void InstallKami(String directory) throws IOException, SAXException, ParserConfigurationException {
-        if (!hasLatestFabric(directory.substring(0, directory.length() - 4))) InstallFabric();
+    private void installKami(String directory) throws IOException, SAXException, ParserConfigurationException {
+        //TODO: find and delete older kami versions
+        if (!hasLatestFabric(directory.substring(0, directory.length() - 5))) installFabric();
 
-        Path mods  = Paths.get(directory + "Kami.jar");
+        String latestVersion = getLatestKami();
+        System.out.println(latestVersion);
 
+        Path mods  = Paths.get(directory + latestVersion + ".jar");
+        InputStream in = new URL("https://github.com/zeroeightysix/KAMI/releases/download/" + latestVersion + "/kami-" + latestVersion + ".jar").openStream();
 
-        //TODO: better download link
-        InputStream in = new URL("https://github.com/zeroeightysix/KAMI/releases/download/1.16.2-aug/kami-1.16.2-aug.jar").openStream();
-        Files.copy(in, mods, StandardCopyOption.REPLACE_EXISTING);
-
-        JOptionPane.showMessageDialog(null, "Installed KAMI to:\n"+directory);
+        if (new File(String.valueOf(mods)).exists()) {
+            JOptionPane.showMessageDialog(null, "It looks like KAMI is already installed.");
+        } else {
+            Files.copy(in, mods, StandardCopyOption.REPLACE_EXISTING);
+            JOptionPane.showMessageDialog(null, "Installed KAMI to:\n"+directory);
+        }
 
         //new File(System.getProperty("user.dir") + File.separator + "fabric-installer.jar").delete();
     }
 
     private boolean hasLatestFabric(String directory) {
-        //TODO: CHECK FOR LATEST VERSION
-        File fabric = new File(directory+"fabric");
-        return fabric.exists();
+        //TODO: CHECK FOR LATEST VERSION OF FABRIC
+        Path fabric = Paths.get(directory+".fabric");
+        System.out.println(fabric.toString());
+        return Files.exists(fabric);
 
     }
 
-    private void InstallFabric() throws IOException, SAXException, ParserConfigurationException {
-        //TODO
-
-
+    private void installFabric() throws IOException, SAXException, ParserConfigurationException {
         URL xmlURL = new URL("https://maven.fabricmc.net/net/fabricmc/fabric-installer/maven-metadata.xml");
         InputStream xml = xmlURL.openStream();
         String latestVersion = DocumentBuilderFactory.newInstance().newDocumentBuilder().parse(xml).getElementsByTagName("latest").item(0).getTextContent();
@@ -160,7 +155,25 @@ public class Installer extends JPanel {
         Files.copy(in, fabricInstall, StandardCopyOption.REPLACE_EXISTING);
 
         Runtime.getRuntime().exec("java -jar fabric-installer.jar");
-        return;
+    }
+
+    private String getLatestKami() throws IOException {
+        try (InputStream is = new URL("https://api.github.com/repos/zeroeightysix/KAMI/releases").openStream()) {
+            BufferedReader reader = new BufferedReader(new InputStreamReader(is, StandardCharsets.UTF_8));
+            StringBuilder builder = new StringBuilder();
+
+            int currentCharacter;
+            while ((currentCharacter = reader.read()) != -1) {
+                builder.append((char) currentCharacter);
+            }
+            String jsonText = builder.toString();
+
+            // As is the case with the Kami Blue installer,
+            // because we can't import libraries that minecraft has and because
+            // the game isn't initialized we **have** to hard code this😔
+            return jsonText.split("tag_name", 2)[1].split("\"", 6)[2];
+
+        }
     }
 
 }
